Replit: Allow position-independent scripts
Pip changes all shebangs in the generated scripts to use the absolute
path of the Python binary in the venv. This works well for most venvs,
but we are not "most venvs". Notably, when repls are forked, the path of
the repl changes, so the new path gets sadly broken.

This change now uses a trick in versions 8.30+ of `env(1)`, which adds
the `-S` flag and can invoke relative paths with ease. Now repls with a
venv can be correctly forked!