Replit: Add a flag to make pip install do nothing
Previously, we were wasting a lot of time actually installing the
packages into a temporary directory, but what we really want is just to
download the artifacts so that the cache is populated.

This change adds a new flag `--noop` to `pip install` that instructs it
to not actually ... you know, install.