Replit: Fix the standalone pip zip (#4)
With the content-addressable cache, `_create_standalone_pip()` would try
to resolve the path of the `__init__.py` file of the `pip` module. Since
that file is a symlink, the path resolution would end up deep inside the
content-addressable cache, causing the generated file to have a single
file with a very unexpected name (the last two components of the sha256
hash of the `__init__.py` file). This meant that if anybody tried to
install any package from source, it would fail miserably because Python
would not be able to find the `pip` module inside the .zip file.

This change skips the resolution part so that pip can create a correct
.zip file with the expected contents.