Replit: Allow more position-independent scripts (#5)
Pip changes all shebangs in the generated scripts to use the absolute
path of the Python binary in the venv. This works well for most venvs,
but we are not "most venvs". Notably, when repls are forked, the path of
the repl changes, so the new path gets sadly broken.

This change now uses the same trick we used in #2 that relies on versions 8.30+
of `env(1)`: call `env` with the add the `-S` flag to invoke relative paths.
Now repls with a venv can be correctly forked!